refactor: expose shared helpers
