+++ src/components/DataTable/index.ts
@@ -1,4 +1,4 @@
-export * from './utils';
+export * from './shared';
 export * from './cells';
 export * from './DataTable';
 export * from './DataTable.types';
