+++ src/components/DataTable/shared/columnHelpers.tsx
@@ -40,4 +40,3 @@ export const dateColumn = <T, >(get: (row: T) => Date, options?: ColumnOptions) 
     sortFn: (a: T, b: T) => get(a).getTime() - get(b).getTime(),
     align: options?.align ?? 'right'
 });
-
