+++ src/components/DataTable/shared/tableHelpers.tsx
@@ -32,12 +32,10 @@ export const createTableColumns = <T extends Record<string, unknown>>(
         cell: ({row, getValue}) => {
             const value = getValue();
 
-            // Use custom render function if provided
             if (col.render) {
                 return col.render(value as T[Exclude<keyof T, SubRowKey>], row.original as T);
             }
 
-            // Fallback: return raw value as ReactNode
             return value as React.ReactNode;
         }
     }));
