+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export {DataTable} from './DataTable';
+export * from './DataTable/shared';
 export * from './Accordion';
 export * from './Accordion/AccordionItem';
 export * from './Badge';
